[New] Created test cases for pages-admin-form-builder-clear-buttons component from enzyme to rtl
